refactor: move folder structure and import location
