+++ src/generator/module/handler/http.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/nurcahyaari/kite/src/ast"
 	"github.com/nurcahyaari/kite/src/generator/protocol"
+	"github.com/nurcahyaari/kite/src/templates"
 	"github.com/nurcahyaari/kite/src/utils/fs"
-	"github.com/nurcahyaari/kite/templates"
 )
 
 type HttpHandlerGen interface {
