+++ src/templates/internaltemplate/loggertemplate/loggertemplate.go
@@ -3,7 +3,7 @@ package loggertemplate
 import (
 	_ "embed"
 
-	"github.com/nurcahyaari/kite/templates"
+	"github.com/nurcahyaari/kite/src/templates"
 )
 
 //go:embed logger.tmpl
