+++ src/templates/internaltemplate/utilstemplate/encryptiontemplate/rsatemplate.go
@@ -3,7 +3,7 @@ package encryptiontemplate
 import (
 	_ "embed"
 
-	"github.com/nurcahyaari/kite/templates"
+	"github.com/nurcahyaari/kite/src/templates"
 )
 
 //go:embed rsa.tmpl
