+++ src/templates/misctemplate/gitignoretemplate.go
@@ -3,7 +3,7 @@ package misctemplate
 import (
 	_ "embed"
 
-	"github.com/nurcahyaari/kite/templates"
+	"github.com/nurcahyaari/kite/src/templates"
 )
 
 //go:embed gitignore.tmpl
