+++ src/templates/misctemplate/maintemplate.go
@@ -3,7 +3,7 @@ package misctemplate
 import (
 	_ "embed"
 
-	"github.com/nurcahyaari/kite/templates"
+	"github.com/nurcahyaari/kite/src/templates"
 )
 
 //go:embed main.tmpl
