+++ src/templates/protocoltemplate/httptemplate/chitemplate/internalhttptemplate.go
@@ -3,7 +3,7 @@ package chitemplate
 import (
 	_ "embed"
 
-	"github.com/nurcahyaari/kite/templates"
+	"github.com/nurcahyaari/kite/src/templates"
 )
 
 //go:embed internalhttptemplate.tmpl
