+++ src/templates/protocoltemplate/httptemplate/chitemplate/responsetemplate.go
@@ -3,7 +3,7 @@ package chitemplate
 import (
 	_ "embed"
 
-	"github.com/nurcahyaari/kite/templates"
+	"github.com/nurcahyaari/kite/src/templates"
 )
 
 //go:embed response.tmpl
